backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic[email]==2.9.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
openai==1.57.0
httpx==0.27.0
apscheduler==3.10.4
resend==2.0.0
alembic==1.13.3